Split vcmiclient in two
Similar to vcmiserver (app) and vcmiservercommon (lib), now
there is vcmiclient (app) and vcmiclientcommon (lib).
